% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmi_plot.R
\name{dmi_plot}
\alias{dmi_plot}
\title{Plot Hybrid Fitness of a Dobzhansky-Muller Incompatibility}
\usage{
dmi_plot(h2, h1, h0)
}
\arguments{
\item{h2}{Fitness of an alternate ancestry homozygous genotype}

\item{h1}{Fitness of a homozygous x heterozygous genotype}

\item{h0}{Fitness of a double heterozygote genotype}
}
\value{
A filled contour plot of fitness across hybrid genotypic space
}
\description{
Plot Hybrid Fitness of a Dobzhansky-Muller Incompatibility
}
