% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{lc_bern}
\alias{lc_bern}
\title{Fit a Bernoulli line cross model}
\usage{
lc_bern(data, ...)
}
\arguments{
\item{data}{A list of inputs for the stan model}

\item{...}{Named arguments to the \code{sample()} method of CmdStan model
objects: \url{https://mc-stan.org/cmdstanr/reference/model-method-sample.html}}
}
\value{
A \code{CmdStanMCMC} object: \url{https://mc-stan.org/cmdstanr/reference/model-method-sample.html}
}
\description{
Fit a Bernoulli line cross model
}
