% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_input}
\alias{prep_input}
\title{Prepare data to go into the Stan model}
\usage{
prep_input(dep_var, anc_prop, anc_het)
}
\arguments{
\item{dep_var}{Dependent variable for the line cross model}

\item{anc_prop}{Ancestry proportion of the individuals in each observation}

\item{anc_het}{Ancestry heterozygosity of the individuals in each observation}
}
\value{
Returns a list containing the data to be input into the Stan program through rstan
}
\description{
Prepare data to go into the Stan model
}
